ci: move from aqora-cli to aqora for pypi
