+++ python/aqora/__init__.py
@@ -1,8 +1,8 @@
 from typing_extensions import Any, TypedDict, override
 
-from . import _aqora_cli
+from . import _aqora
 from .notebook import notebook
-from aqora_cli._aqora_cli import *  # pyright: ignore[reportAssignmentType, reportWildcardImportFromLibrary]  # noqa: F403
+from aqora._aqora import *  # pyright: ignore[reportAssignmentType, reportWildcardImportFromLibrary]  # noqa: F403
 
 
 class GraphQLError(TypedDict):
@@ -29,4 +29,4 @@ class ClientError(Exception):
         return self.message
 
 
-__doc__ = _aqora_cli.__doc__
+__doc__ = _aqora.__doc__
