+++ template/assets/dataset_marimo/pyproject.toml.hbs
@@ -5,7 +5,7 @@ requires-python = ">={{ python_version }}"
 
 dependencies = [
   "marimo[recommended,lsp]>={{ marimo_version }}",
-  "aqora-cli[pyarrow]>={{ cli_version }}",
+  "aqora[pyarrow]>={{ cli_version }}",
 ]
 
 [dependency-groups]
