+++ template/assets/use_case/pyproject.toml.hbs
@@ -3,7 +3,7 @@ name = "use-case"
 version = "0.1.0"
 requires-python = ">={{ python_version }}"
 
-dependencies = ["aqora-cli>={{ cli_version }}"]
+dependencies = ["aqora>={{ cli_version }}"]
 
 [build-system]
 requires = [
